Move transaction tests to transaction package
